LPS-63470 rename module
