--- modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-action-executor/build.gradle
@@ -1,12 +1,0 @@
-dependencies {
-	provided group: "com.liferay.portal", name: "portal-service", version: liferay.portalVersion
-	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:business-productivity:portal-workflow:portal-workflow-kaleo-api")
-	provided project(":apps:business-productivity:portal-workflow:portal-workflow-kaleo-service")
-	provided project(":apps:platform:osgi-service-tracker-collections")
-	provided project(":apps:platform:portal-rules-engine:portal-rules-engine-api")
-}
-
-liferay {
-	deployDir = file("${liferayHome}/osgi/portal")
-}
