+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-runtime-action-executor/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Portal Workflow Kaleo Runtime Action Executor
+Bundle-SymbolicName: com.liferay.portal.workflow.kaleo.runtime.action.executor
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.workflow.kaleo.runtime.action.executor
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Kaleo Workflow Engine
+-include: ../../../marketplace/business-productivity/bnd.bnd
