+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-runtime-action-executor/build.gradle
@@ -1,0 +1,12 @@
+dependencies {
+	provided group: "com.liferay.portal", name: "portal-service", version: liferay.portalVersion
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":core:osgi-service-tracker-collections")
+	provided project(":portal-rules-engine:portal-rules-engine-api")
+	provided project(":portal-workflow:portal-workflow-kaleo-api")
+	provided project(":portal-workflow:portal-workflow-kaleo-service")
+}
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/portal")
+}
