+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-runtime-action-executor/src/main/java/com/liferay/portal/workflow/kaleo/runtime/action/executor/ActionExecutor.java
@@ -12,28 +12,18 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.action.executor;
+package com.liferay.portal.workflow.kaleo.runtime.action.executor;
 
-import com.liferay.portal.kernel.workflow.WorkflowException;
+import com.liferay.portal.workflow.kaleo.model.KaleoAction;
+import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
 
 /**
  * @author Michael C. Han
  */
-public class ActionExecutorException extends WorkflowException {
+public interface ActionExecutor {
 
-	public ActionExecutorException() {
-	}
-
-	public ActionExecutorException(String msg) {
-		super(msg);
-	}
-
-	public ActionExecutorException(String msg, Throwable cause) {
-		super(msg, cause);
-	}
-
-	public ActionExecutorException(Throwable cause) {
-		super(cause);
-	}
+	public void execute(
+			KaleoAction kaleoAction, ExecutionContext executionContext)
+		throws ActionExecutorException;
 
 }
