+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-runtime-action-executor/src/main/java/com/liferay/portal/workflow/kaleo/runtime/action/executor/ActionExecutorException.java
@@ -12,18 +12,28 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.action.executor;
+package com.liferay.portal.workflow.kaleo.runtime.action.executor;
 
-import com.liferay.portal.workflow.kaleo.model.KaleoAction;
-import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
+import com.liferay.portal.kernel.workflow.WorkflowException;
 
 /**
  * @author Michael C. Han
  */
-public interface ActionExecutor {
+public class ActionExecutorException extends WorkflowException {
 
-	public void execute(
-			KaleoAction kaleoAction, ExecutionContext executionContext)
-		throws ActionExecutorException;
+	public ActionExecutorException() {
+	}
+
+	public ActionExecutorException(String msg) {
+		super(msg);
+	}
+
+	public ActionExecutorException(String msg, Throwable cause) {
+		super(msg, cause);
+	}
+
+	public ActionExecutorException(Throwable cause) {
+		super(cause);
+	}
 
 }
