+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-runtime-action-executor/src/main/java/com/liferay/portal/workflow/kaleo/runtime/action/executor/internal/DRLActionExecutor.java
@@ -12,15 +12,15 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.action.executor.internal;
+package com.liferay.portal.workflow.kaleo.runtime.action.executor.internal;
 
 import com.liferay.portal.kernel.resource.StringResourceRetriever;
 import com.liferay.portal.rules.engine.Fact;
 import com.liferay.portal.rules.engine.Query;
 import com.liferay.portal.rules.engine.RulesEngine;
 import com.liferay.portal.rules.engine.RulesResourceRetriever;
-import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutor;
-import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutorException;
+import com.liferay.portal.workflow.kaleo.runtime.action.executor.ActionExecutor;
+import com.liferay.portal.workflow.kaleo.runtime.action.executor.ActionExecutorException;
 import com.liferay.portal.workflow.kaleo.model.KaleoAction;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
 import com.liferay.portal.workflow.kaleo.runtime.util.RulesContextBuilder;
@@ -39,7 +39,7 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true,
-	property = {"com.liferay.portal.workflow.kaleo.action.executor.language=drl"},
+	property = {"com.liferay.portal.workflow.kaleo.runtime.action.executor.language=drl"},
 	service = ActionExecutor.class
 )
 public class DRLActionExecutor implements ActionExecutor {
